@@ -1,10 +1,17 @@
 import log from 'electron-log'
 
 import { NATIVE_CURRENCY } from '../../../resources/constants'
+import { toTokenId } from '../../../resources/domain/balance'
 import BalancesWorkerController from './controller'
 import { CurrencyBalance, TokenBalance } from './scan'
 
 const RESTART_WAIT = 5 // seconds
+
+// time to wait in between scans, in seconds
+const scanInterval = {
+  active: 20,
+  inactive: 60 * 10
+}
 
 export default function (store: Store) {
   const storeApi = {
@@ -17,7 +24,7 @@ export default function (store: Store) {
         .filter(n => (n.connection.primary || {}).connected || (n.connection.secondary || {}).connected)
     },
     getCustomTokens: () => (store('main.tokens.custom') || []) as Token[],
-    getKnownTokens: (address?: Address) => ((address && store('main.tokens.known', address)) || []) as Token[],
+    getKnownTokens: (address?: Address): Token[] => (address && store('main.tokens.known', address)) || [],
     getCurrencyBalances: (address: Address) => {
       return ((store('main.balances', address) || []) as Balance[])
         .filter(balance => balance.address === NATIVE_CURRENCY)
@@ -25,11 +32,12 @@ export default function (store: Store) {
     getTokenBalances: (address: Address) => {
       return ((store('main.balances', address) || []) as Balance[])
         .filter(balance => balance.address !== NATIVE_CURRENCY)
-    }
+    },
   }
 
   let scan: NodeJS.Timeout | null
   let workerController: BalancesWorkerController | null
+  let onResume: Function | null
 
   function attemptRestart () {
     log.warn(`balances controller stopped, restarting in ${RESTART_WAIT} seconds`)
@@ -70,11 +78,38 @@ export default function (store: Store) {
     workerController.on('tokenBalances', (address, balances) => {
       handleUpdate(address, handleTokenBalanceUpdate.bind(null, balances))
     })
+
+    workerController.on('tokenBlacklist', (address, tokens) => {
+      handleUpdate(address, handleTokenBlacklistUpdate.bind(null, tokens))
+    })
   }
 
   function restart () {
     start()
     setAddress(storeApi.getActiveAddress())
+  }
+
+  function resume () {
+    if (onResume) onResume()
+
+    onResume = null
+  }
+
+  function pause () {
+    if (stopScan()) {
+      log.debug('Pausing balances scan')
+
+      const address = storeApi.getActiveAddress()
+
+      // even when paused ensure data is updated every 10 minutes
+      resetScan(address, scanInterval.inactive)
+
+      onResume = () => {
+        log.debug(`Resuming balances scan for address ${address}`)
+
+        startScan(address)
+      }
+    }
   }
 
   function stop () {
@@ -93,29 +128,43 @@ export default function (store: Store) {
   function startScan (address: Address) {
     stopScan()
 
-    log.verbose(`starting balances scan for ${address}`)
+    if (onResume) onResume = null
 
-    const scanForAddress = () => {
-      // update balances for the active account every 20 seconds
+    log.verbose(`Starting balances scan for ${address}`)
+
+    const initiateScan = () => {
+      // do an initial scan before starting the timer
       setTimeout(() => {
         updateActiveBalances(address)
       }, 0)
 
-      scan = setTimeout(() => {
-        if (workerController?.isRunning()) scanForAddress()
-      }, 20 * 1000)
+      resetScan(address, scanInterval.active)
     }
 
-    runWhenReady(() => scanForAddress())
+    runWhenReady(() => initiateScan())
   }
 
   function stopScan () {
-    log.debug('stopping balances scan')
-
     if (scan) {
       clearTimeout(scan)
       scan = null
+
+      return true
     }
+
+    return false
+  }
+
+  function resetScan (address: Address, interval: number) {
+    scan = setTimeout(() => {
+      if (workerController?.isRunning()) {
+        setTimeout(() => {
+          updateActiveBalances(address)
+        }, 0)
+      }
+
+      resetScan(address, interval)
+    }, interval * 1000)
   }
 
   function updateActiveBalances (address: Address) {
@@ -178,21 +227,26 @@ export default function (store: Store) {
     if (changedBalances.length > 0) {
       store.setBalances(address, changedBalances)
 
-      const knownTokens = storeApi.getKnownTokens(address)
+      const knownTokens = new Set(storeApi.getKnownTokens(address).map(toTokenId))
+      const isKnown = (balance: TokenBalance) => knownTokens.has(toTokenId(balance))
 
       // add any non-zero balances to the list of known tokens
-      const unknownBalances = changedBalances
-        .filter(b => parseInt(b.balance) > 0 && !knownTokens.some(t => t.address === b.address && t.chainId === b.chainId))
+      const unknownBalances = changedBalances.filter(b => parseInt(b.balance) > 0 && !isKnown(b))
 
       if (unknownBalances.length > 0) {
         store.addKnownTokens(address, unknownBalances)
       }
 
       // remove zero balances from the list of known tokens
-      const zeroBalances = changedBalances
-        .filter(b => parseInt(b.balance) === 0 && knownTokens.some(t => t.address === b.address && t.chainId === b.chainId))
+      const zeroBalances = changedBalances.reduce((zeroBalSet, balance) => {
+        const tokenId = toTokenId(balance)
+        if (parseInt(balance.balance) === 0 && knownTokens.has(tokenId)) {
+          zeroBalSet.add(tokenId)
+        }
+        return zeroBalSet
+      }, new Set<string>())
 
-      if (zeroBalances.length > 0) {
+      if (zeroBalances.size) {
         store.removeKnownTokens(address, zeroBalances)
       }
     }
@@ -200,12 +254,31 @@ export default function (store: Store) {
     store.accountTokensUpdated(address)
   }
 
+  function handleTokenBlacklistUpdate (tokensToRemove: Set<string>) {
+    const includesBlacklistedTokens = (arr: WithTokenId[]) => arr.some(val => tokensToRemove.has(toTokenId(val)))
+
+    const balances: Record<string, Balance[]> = store('main.balances')
+    const knownTokens: Record<string, Token[]> = store('main.tokens.known')
+
+    Object.entries(balances).forEach(([accountAddress, balances]) => {
+      if(includesBlacklistedTokens(balances)){
+        store.removeBalances(accountAddress, tokensToRemove)
+      }
+    })
+
+    Object.entries(knownTokens).forEach(([accountAddress, tokens]) => {
+      if(includesBlacklistedTokens(tokens)){
+        store.removeKnownTokens(accountAddress, tokensToRemove)
+      }
+    })
+  }
+
   function setAddress (address: Address) {
     if (!workerController) {
       log.warn(`tried to set address to ${address} but balances controller is not running`)
       return
     }
-
+ 
     if (address) {
       log.verbose('setting address for balances updates', address)
       startScan(address)
@@ -235,5 +308,5 @@ export default function (store: Store) {
     runWhenReady(() => workerController?.updateKnownTokenBalances(address, tokens))
   }
 
-  return { start, stop, setAddress, addNetworks, addTokens }
+  return { start, stop, resume, pause, setAddress, addNetworks, addTokens }
 }
