@@ -1,7 +1,6 @@
 import log from 'electron-log'
 
 import ethProvider from 'eth-provider'
-import sushiswapTokenList from '@sushiswap/default-token-list'
 
 import nebulaApi from '../../nebula'
 import defaultTokenList from './default-tokens.json'
@@ -9,60 +8,32 @@ import defaultTokenList from './default-tokens.json'
 const TOKENS_ENS_DOMAIN = 'tokens.frame.eth'
 
 interface TokenSpec extends Token {
-  extensions: {
-    omit: boolean
+  extensions?: {
+    omit?: boolean
   }
 }
 
-function mergeTokens (existingTokens: Token[], updatedTokens: TokenSpec[]) {
-  const omitList: string[] = []
-
-  const mergedList = [existingTokens, updatedTokens].reduce((tokens, list) => {
-    list.forEach(token => {
-      const address = token.address.toLowerCase()
-      const key = `${token.chainId}:${address}`
-      const omitToken = ((token as any).extensions || {}).omit
-
-      if (omitToken) {
-        omitList.push(key)
-        delete tokens[key]
-      } else if (!omitList.includes(key) && !(key in tokens)) {
-        tokens[key] = { ...token, address }
-      }
-    })
-
-    return tokens
-  }, {} as { [key: string]: Token })
-
-  return Object.values(mergedList)
+function isBlacklisted (token: TokenSpec) {
+  return token.extensions?.omit
 }
 
 export default class TokenLoader {
-  private tokenList: Token[] = []
+  private tokens: TokenSpec[] = defaultTokenList.tokens as TokenSpec[]
   private nextLoad?: NodeJS.Timeout | null
 
   private readonly eth = ethProvider('frame', { origin: 'frame-internal', name: 'tokenLoader' })
   private readonly nebula = nebulaApi(this.eth)
 
   constructor () {
-    // token resolution uses mainnet ENS
     this.eth.setChain('0x1')
-
-    this.tokenList = mergeTokens(
-      sushiswapTokenList.tokens as Token[],
-      defaultTokenList.tokens as TokenSpec[]
-    )
   }
 
   private async loadTokenList (timeout = 60_000) {
     try {
       const updatedTokens = await this.fetchTokenList(timeout)
-
       log.info(`Fetched ${updatedTokens.length} tokens`)
-
-      this.tokenList = mergeTokens(this.tokenList, updatedTokens)
-
-      log.info(`Updated token list to contain ${this.tokenList.length} tokens`)
+      this.tokens = updatedTokens
+      log.info(`Updated token list to contain ${this.tokens.length} tokens`)
 
       this.nextLoad = setTimeout(() => this.loadTokenList(), 10 * 60_000)
     } catch (e) {
@@ -131,7 +102,13 @@ export default class TokenLoader {
     }
   }
 
-  getTokens (chainId: number) {
-    return this.tokenList.filter(token => token.chainId === chainId)
+  getTokens (chains: number[]) {
+    return this.tokens.filter(token => !isBlacklisted(token) && chains.includes(token.chainId))
+  }
+
+  getBlacklist (chains: number[] = []) {
+    const chainMatches = (token: TokenSpec) => !chains.length || chains.includes(token.chainId)
+
+    return this.tokens.filter((token) => isBlacklisted(token) && chainMatches(token))
   }
 }
