@@ -1,10 +1,9 @@
 import { app, ipcMain, protocol, shell, clipboard, globalShortcut, powerMonitor, BrowserWindow } from 'electron'
 import path from 'path'
 import log from 'electron-log'
-import { numberToHex } from 'web3-utils'
 import url from 'url'
 
-// DO NOT MOVE - env var below is required to enable watch mode for development on the renderer process and must be set before all local imports 
+// DO NOT MOVE - env var below is required for app init and must be set before all local imports 
 process.env.BUNDLE_LOCATION = process.env.BUNDLE_LOCATION || path.resolve(__dirname, './../..', 'bundle')
 
 import * as errors from './errors'
@@ -19,7 +18,6 @@ import signers from './signers'
 import persist from './store/persist'
 import showUnhandledExceptionDialog from './windows/dialog/unhandledException'
 import Erc20Contract from './contracts/erc20'
-import provider from './provider'
 import { getErrorCode } from '../resources/utils'
 import { FrameInstance } from './windows/frames/frameInstances'
 
@@ -30,15 +28,34 @@ app.commandLine.appendSwitch('ignore-gpu-blacklist', 'true')
 app.commandLine.appendSwitch('enable-native-gpu-memory-buffers', 'true')
 app.commandLine.appendSwitch('force-color-profile', 'srgb')
 
-log.transports.console.level = process.env.LOG_LEVEL || 'info'
+const isDev = process.env.NODE_ENV === 'development'
+
+log.transports.console.level = process.env.LOG_LEVEL || (isDev ? 'verbose' : 'info')
 log.transports.file.level = ['development', 'test'].includes(process.env.NODE_ENV) ? false : 'verbose'
 
-const dev = process.env.NODE_ENV === 'development'
 const hasInstanceLock = app.requestSingleInstanceLock()
 
 if (!hasInstanceLock) {
   log.info('another instance of Frame is running - exiting...')
   app.exit(1)
+}
+
+if (isDev) {
+  const cpuMonitoringInterval = 10 // seconds
+  const cpuThreshold = 30 // percent
+
+  setTimeout(() => {
+    app.getAppMetrics()
+
+    setInterval(() => {
+      const cpuUsers = app.getAppMetrics().filter(metric => metric.cpu.percentCPUUsage > cpuThreshold)
+
+      if (cpuUsers.length > 0) {
+        log.verbose(`Following processes used more than ${cpuThreshold}% CPU over the last ${cpuMonitoringInterval} seconds`)
+        log.verbose(JSON.stringify(cpuUsers, undefined, 2))
+      }
+    }, cpuMonitoringInterval * 1000)
+  }, 10_000)
 }
 
 require('./rpc')
@@ -256,7 +273,7 @@ ipcMain.on('tray:syncPath', (e, path, value) => {
 ipcMain.on('tray:ready', () => {
   require('./api')
 
-  if (!dev) {
+  if (!isDev) {
     startUpdater()
   }
 })
