@@ -10,9 +10,9 @@ const nav = {
     // Adds new crumb to nav array
     store.navForward(windowId, crumb)
   },
-  back: (windowId: string) => {
+  back: (windowId: string, steps = 1) => {
     // Removes last crumb from nav array
-    store.navBack(windowId)
+    store.navBack(windowId, steps)
   },
   update: (windowId: string, crumb: Breadcrumb, navigate: boolean = true) => {
     // Updated last crumb in nav array with new data
@@ -26,8 +26,8 @@ ipcMain.on('nav:forward', (e, windowId: string, crumb: Breadcrumb) => {
   nav.forward(windowId, crumb)
 })
 
-ipcMain.on('nav:back', (e, windowId: string) => {
-  nav.back(windowId)
+ipcMain.on('nav:back', (e, windowId: string, steps = 1) => {
+  nav.back(windowId, steps)
 })
 
 ipcMain.on('nav:update', (e, windowId: string, crumb: Breadcrumb, navigate: boolean) => {
